Use new proposal api
