+++ package.json
@@ -824,5 +824,7 @@
         "ms-vscode.azure-account",
         "ms-azuretools.vscode-azureresourcegroups"
     ],
-    "enableProposedApi": true
+    "enabledApiProposals": [
+        "terminalDataWriteEvent"
+    ]
 }
